// activity-points/src/app/useroptions/interfaces.ts
export interface Cat12FinalData {
  level: string;
  prize: string;
}

export interface Cat3FinalData {
  specification: string;
  level: string;
  setLevel(lev: string);
}

export interface Cat5FinalData {
  specification: string;
}

export interface Cat6FinalData {
  specification: string;
  role: string;
}